Fix custom port assignment
Fixes the ability to use custom ports for rabbitmq: the underlying
march hare library disregards the value of the port parameter
if the :hosts option is set to provide multiple addresses. This
commit changes the setup to use the :addresses option, which
enables a port to be included.

Fixes #53

Fixes #54