output: propagate back-pressure when connection is blocked
When the other end of a RabbitMQ connection is either unwilling or unable to
continue reading bytes from its underlying TCP stream, the connection is
flagged as "blocked", but attempts to publish onto exchanges using the
connection will not block in the client.

Here we hook into notifications of connection-blocked state to set up a
`BackPressure::GatedExecutor`, which is used to prevent runaway writes when
publishing to an exchange on a blocked connection.