Doc:Replace plugin_header file with plugin_header-integration file
Integration plugins need a header that is different from the stand-alone plugin
header. The plugin docs links should point to the integration repo rather than the
stand-alone input and output repos. The header also adds a note that the plugin
is part of an integration rather than stand-alone.

The new header included in this work was added with Logstash #11891.